server/requirements.txt
Django==4.2.23
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-rest-passwordreset==1.2.1
python-decouple==3.8
sqlparse==0.5.3
asgiref==3.8.1
tzdata==2025.2
typing-extensions==4.13.2
django-cors-headers==4.9.0
Pillow==11.3.0
python-dotenv==1.0.1
django-filter==24.2
cloudinary==1.40.0
django-allauth==0.54.0
requests==2.31.0
gunicorn
